// src/app/api/shared/models/abv/insurance-view.model.ts
export interface InsuranceViewModel {
    insurancePlanType: string;
    pbmName: string;
    pbmPhone: string;
    policyIDNumber: string;
    groupNumber: string;
    binNumber: string;
    pcnNumber: string;
    policyholderRelationshipToPatient: string;
    policyholderDateOfBirth: string;
    policyholderFirstName: string;
    policyholderLastName: string;
}